🧪 Add dedicated tests for SafeStorage helper
Extracted SafeStorage tests from utils.test.js into a dedicated storage.test.js file and ensured full test coverage for success, mocked failures, and silent error catching behaviors.
